@@ -1,22 +1,21 @@
 package org.woodwhales.music.config;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.ProviderManager;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.woodwhales.music.security.MyAuthenticationFailureHandler;
-import org.woodwhales.music.security.TwoFactorAuthenticationSuccessHandler;
+import org.woodwhales.music.security.MyAuthenticationSuccessHandler;
 
 /**
  * spring security 配置文件
@@ -25,79 +24,60 @@ import org.woodwhales.music.security.TwoFactorAuthenticationSuccessHandler;
 @Slf4j
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(securedEnabled = true)
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-
-	@Autowired
-	private UserDetailsService userDetailsService;
-
-	@Autowired
-	private MyAuthenticationFailureHandler myAuthenticationFailureHandler;
+public class WebSecurityConfig {
 
 	@Bean
 	public PasswordEncoder passwordEncoder() {
 		return new BCryptPasswordEncoder();
 	}
 
+	@Bean
+	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+		// 禁用basic明文验证
+		http.httpBasic(Customizer.withDefaults());
+		// 前后端分离架构不需要csrf保护
+		http.csrf(AbstractHttpConfigurer::disable);
+		http.authorizeHttpRequests(authorize -> {
+			authorize.requestMatchers("/", "/index", "/error").permitAll()
+					.requestMatchers(
+							"/admin/fonts/**",
+							"/admin/css/**",
+							"/admin/images/**",
+							"/admin/js/**",
+							"/admin/lib/**",
+							"/webjars/**",
+							"/images/**",
+							"/css/**",
+							"/fonts/**",
+							"/js/**",
+							"/**.png",
+							"/**.ico",
+							"/favicon.ico").permitAll()
+					.anyRequest().authenticated();
+		});
 
-	@Override
-	public void configure(HttpSecurity http) throws Exception {
-		http.formLogin() 								// 定义当需要用户登录时候，转到的登录页面。
-				.loginPage("/admin/login")	 					// 设置登录页面
-				.loginProcessingUrl("/admin/loginTo") 			// 自定义的登录接口，如果没有配置，则使用：loginPage
-				//.successForwardUrl()
-				.successHandler(new TwoFactorAuthenticationSuccessHandler())
-
-				//.defaultSuccessUrl("/admin/")					// login 成功之后，默认跳转的页面
-				// .failureForwardUrl() // 本质是添加了一个 链接重定向处理器（failureHandler）： = org.springframework.security.web.authentication.ForwardAuthenticationFailureHandler
-				.failureHandler(myAuthenticationFailureHandler)
-				.and()
-				// ========== 登出配置 ==========
-				.logout()
-				.logoutUrl("/admin/logout")
-				.logoutSuccessUrl("/admin/login")  				// logout 成功之后，默认跳转的页面
-				.permitAll()
-				.and()
-
-				// ========== 定义哪些URL需要被保护、哪些不需要被保护 ==========
-				.authorizeRequests()
-				.antMatchers("/", "/index", "/admin/login", "/admin/two-factor")
-				.permitAll()
-
-				.anyRequest()// 设置所有人都可以访问登录页面
-				.authenticated() 				// 任何请求,登录后可以访问
-
-				.and()
-				.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-
-				.and()
-				.csrf()
-				.disable();
-		;
+		http.formLogin(formLogin ->
+			formLogin.loginPage("/admin/login").permitAll()
+					.loginProcessingUrl("/admin/loginTo").permitAll()
+					.successHandler(new MyAuthenticationSuccessHandler("/admin/"))
+					.failureHandler(new MyAuthenticationFailureHandler("/admin/login"))
+		);
+		http.logout(formLogout ->
+				formLogout.logoutUrl("/logout").permitAll()
+						.logoutSuccessUrl("/login").permitAll()
+						.invalidateHttpSession(true));
+		http.securityContext(securityContext -> securityContext.requireExplicitSave(false));
+		return http.build();
 	}
 
-	@Override
-	public void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.userDetailsService(userDetailsService)
-				.passwordEncoder(passwordEncoder());
-	 }
-
-	@Override
-    public void configure(WebSecurity web) throws Exception {
-        // 设置拦截忽略文件夹，可以对静态资源放行
-        web.ignoring().antMatchers("/admin/fonts/**",
-				"/admin/css/**",
-				"/admin/images/**",
-				"/admin/js/**",
-				"/admin/lib/**",
-				"/webjars/**",
-				"/images/**",
-				"/css/**",
-				"/fonts/**",
-				"/js/**",
-				"/**.png",
-				"/**.ico",
-				"/favicon.ico");
-    }
+	@Bean
+	public AuthenticationManager authenticationManager(
+			UserDetailsService userDetailsService,
+			PasswordEncoder passwordEncoder) {
+		DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
+		authenticationProvider.setUserDetailsService(userDetailsService);
+		authenticationProvider.setPasswordEncoder(passwordEncoder);
+		return new ProviderManager(authenticationProvider);
+	}
 
 }
