@@ -1,10 +1,9 @@
 package org.woodwhales.music.controller.param;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
 
 /**
  * @author woodwhales on 2024-05-09 10:11
