@@ -1,15 +1,14 @@
 package org.woodwhales.music.security;
 
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
 import org.springframework.security.web.util.UrlUtils;
-import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
@@ -17,8 +16,10 @@ import java.io.IOException;
  * @author woodwhales
  */
 @Slf4j
-@Component("myAuthenticationFailureHandler")
 public class MyAuthenticationFailureHandler extends SimpleUrlAuthenticationFailureHandler {
+
+    private final String forwardUrl;
+    
 
     @Override
     public void onAuthenticationFailure(HttpServletRequest request,
@@ -29,9 +30,9 @@ public class MyAuthenticationFailureHandler extends SimpleUrlAuthenticationFailu
         this.getRedirectStrategy().sendRedirect(request, response, "/admin/login");
     }
 
-    public MyAuthenticationFailureHandler forwardUrl(String forwardUrl) {
+    public MyAuthenticationFailureHandler(String forwardUrl) {
         Assert.isTrue(UrlUtils.isValidRedirectUrl(forwardUrl), () -> "'" + forwardUrl + "' is not a valid forward URL");
-        return this;
+        this.forwardUrl = forwardUrl;
     }
 
 }
